Add ImGui docking. Make Bubble static lib.

+++ Bubble/src/Platform/Windows/WindowsInput.cpp
@@ -30,7 +30,7 @@ namespace bubble
 
 		glfwGetCursorPos(window, &xPos, &yPos);
 
-		return { xPos, yPos };
+		return { static_cast<float>(xPos), static_cast<float>(yPos) };
 	}
 
 	float WindowsInput::getMouseXImpl()
